requirements_server.txt
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.3.1
python-dotenv==1.0.0
pytz==2024.1
